styling

+++ src/main/java/com/example/springzhdan/repository/ReviewRepository.java
@@ -14,5 +14,5 @@ public interface ReviewRepository extends JpaRepository<Review, Long> {
     List<Review> check (boolean product);
 
     @Query("select avg(r.rate) from Review r where r.product.id=:product and r.published=true")
-    Integer rate (long product);
+    Double rate (long product);
 }
